+++ fastly/resource_fastly_domain_v1_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	gofastly "github.com/fastly/go-fastly/v9/fastly"
-	v1 "github.com/fastly/go-fastly/v9/fastly/domains/v1"
+	gofastly "github.com/fastly/go-fastly/v10/fastly"
+	v1 "github.com/fastly/go-fastly/v10/fastly/domains/v1"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"