+++ fastly/resource_fastly_object_storage_access_key_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"text/template"
 
-	gofastly "github.com/fastly/go-fastly/v9/fastly"
-	"github.com/fastly/go-fastly/v9/fastly/objectstorage/accesskeys"
+	gofastly "github.com/fastly/go-fastly/v10/fastly"
+	"github.com/fastly/go-fastly/v10/fastly/objectstorage/accesskeys"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"